+++ omega_gamecode/code/cgame/cg_local.h
@@ -1218,6 +1218,8 @@ typedef struct {
 	int				redflag, blueflag;		// flag status from configstrings
 	int				flagStatus;
 
+	int				redLivingCount, blueLivingCount;
+
 	qboolean  newHud;
 
 	//
@@ -1478,7 +1480,6 @@ extern vmCvar_t cg_drawFriendThroughWalls;
 extern vmCvar_t	cg_drawItemPickup;
 extern vmCvar_t	cg_drawTeamBackground;
 extern vmCvar_t cg_elimination_activewarmup;
-extern vmCvar_t cg_elimination_allgametypes;
 extern vmCvar_t cg_enemyColor;
 extern vmCvar_t cg_enemyModel;
 extern vmCvar_t cg_enemySounds;