+++ omega_gamecode/code/game/g_combat.c
@@ -912,6 +912,9 @@ void player_die( gentity_t *self, gentity_t *inflictor, gentity_t *attacker, int
 
 	trap_LinkEntity (self);
 
+	if ( G_IsElimTeamGametype() ) {
+		CheckTeamCount();
+	}
 }